# swagger.yml
openapi: "3.0.3"
info:
  title: Movie Theather Web Development Project
  version: "1.0"
servers:
  - url: http://localhost:3001/
paths:
  /user/login:
    post:
      summary: Realizar login y obtener clave de autorización
      tags:
        - User
      requestBody:
        description: Required
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_name:
                  type: string
                  example: admin
                pwd:
                  type: string
                  example: "123"
              required:
                - user_name
                - pwd
      responses:
        "200":
          description: OK
        "404":
          description: Usuario no encontrado
  /user:
    get:
      summary: Obtener usuario decodificado
      tags:
        - User
      security:
        - JWT:
            [
              eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6OSwidXNlcl9uYW1lIjoiYWRtaW4iLCJlbWFpbCI6Im1hcmxpZXJhbWlyZXphQGdtYWlsLmNvbSIsInRva2VuIjoiJDJiJDEwJE5aWnNxc1ZPSlA2Zzh5Q0hqM1BmNC5ZQW5mdElUNHFWeUpEd29IaEhKRW8vNDljekpoY25tIiwiY3JlYXRlZF9hdCI6IjIwMjUtMDMtMTVUMDY6MDA6MDAuMDAwWiIsInJvbGUiOiJhZG1pbiIsImVzdGFkbyI6ImFjdGl2ZSIsImlhdCI6MTc0MzA0MzUxNX0.kuxHJzRxR3q8Q8TZ_r4tg-8GJv23HG7WIvCtY_rpYsA,
            ]

      responses:
        "200":
          description: OK
        "404":
          description: Usuario no encontrado
  /user/new:
    post:
      tags:
        - User
      summary: Crear una cuenta de cliente
      requestBody:
        description: Atributos de un cliente
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_name:
                  type: string
                email:
                  type: string
                pwd:
                  type: string
              required:
                - user_name
                - pwd
      responses:
        "200":
          description: OK

  /admin/cinema:
    post:
      security:
        - JWT: []

      summary: Crear los cines y horarios para la pelicula
      tags:
        - Cinema - Admin
      requestBody:
        description: Cuerpo de request de cinemas
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                rows:
                  type: number
                columns:
                  type: number
                movie:
                  type: string
                img:
                  type: string
              required:
                - name
                - rows
                - columns
                - movie
                - img
      responses:
        "200":
          description: OK
        "401":
          description: Acceso denegado
  /admin/cinema/{id}:
    put:
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true

      security:
        - JWT: []
      summary: Modificar la capacidad de una sala
      tags:
        - Cinema - Admin
      requestBody:
        description: Cuerpo de request de cinemas
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rows:
                  type: number
                columns:
                  type: number
              required:
                - rows
                - columns
      responses:
        "200":
          description: OK
        "401":
          description: Acceso denegado
  /admin/movies/{id}:
    put:
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true

      security:
        - JWT: []
      summary: Modificar la información de una pelicula
      tags:
        - Cinema - Admin
      requestBody:
        description: Cuerpo de request de cinemas
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                movie:
                  type: string
                img:
                  type: string
              required:
                - movie
                - img
      responses:
        "200":
          description: OK
        "401":
          description: Acceso denegado
  /admin/user/{id}:
    delete:
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true

      security:
        - JWT: []
      summary: Dar de baja a un usuario
      tags:
        - User
      responses:
        "200":
          description: OK
        "401":
          description: Acceso denegado

  /cinema:
    get:
      summary: Obtener todas las salas a futuro
      tags:
        - Cinema
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
  /cinema/{id}:
    get:
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      summary: Obtener datos de sala
      tags:
        - Cinema
      responses:
        "200":
          description: OK
  /schedule:
    get:
      summary: Obtener todas las funciones de una sala
      tags:
        - Cinema
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
  /seats:
    get:
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
      summary: Obtener la disponibilidad de una funcion
      tags:
        - Cinema

      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Acceso denegado
    post:
      security:
        - JWT: []

      summary: Relaizar reserva
      tags:
        - Cinema
      requestBody:
        description: Cuerpo de request de reserva
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                column:
                  type: integer
                rows:
                  type: integer
                id_user:
                  type: integer
                id_schedule:
                  type: integer
              required:
                - name
                - rows
                - columns
                - id_user
                - id_schedule
      responses:
        "200":
          description: OK
        "401":
          description: Acceso denegado
        "400":
          description: Bad request

components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      #eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6OSwidXNlcl9uYW1lIjoiYWRtaW4iLCJlbWFpbCI6Im1hcmxpZXJhbWlyZXphQGdtYWlsLmNvbSIsInRva2VuIjoiJDJiJDEwJE5aWnNxc1ZPSlA2Zzh5Q0hqM1BmNC5ZQW5mdElUNHFWeUpEd29IaEhKRW8vNDljekpoY25tIiwiY3JlYXRlZF9hdCI6IjIwMjUtMDMtMTVUMDY6MDA6MDAuMDAwWiIsInJvbGUiOiJhZG1pbiIsImVzdGFkbyI6ImFjdGl2ZSIsImlhdCI6MTc0MzA0MzUxNX0.kuxHJzRxR3q8Q8TZ_r4tg-8GJv23HG7WIvCtY_rpYsA